+++ .gitignore
@@ -1,2 +1,4 @@
 *.csv
 *.yag
+*.zrd
+*.xlsx
