--- jdeu_start_v2.ps1
@@ -1,8 +1,0 @@
-# Get the directory of the currently executing script
-$scriptDirectory = $PSScriptRoot
-
-# Change to the script directory
-cd $scriptDirectory
-
-# Run the Python script
-python ./jdeu_gui_module.py
