--- nekot.zrd
@@ -1,3 +1,0 @@
-MTI0MjQyOTY3NTkwOjx9McJg/Do809z4qkgrYr0wY5RQ
-
-git push https://verdictzero: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@github.com/verdictzero/jira_experiment
